Rollback possible breaking change
