+++ router/src/test/groovy/io/micronaut/web/router/resource/StaticResourceResolverSpec.groovy
@@ -16,8 +16,6 @@
 package io.micronaut.web.router.resource
 
 import io.micronaut.core.io.ResourceResolver
-import io.micronaut.web.router.resource.StaticResourceConfiguration
-import io.micronaut.web.router.resource.DefaultStaticResourceResolver
 import spock.lang.Specification
 
 class StaticResourceResolverSpec extends Specification {
@@ -31,7 +29,7 @@ class StaticResourceResolverSpec extends Specification {
         StaticResourceConfiguration config2 = new StaticResourceConfiguration(rr, null)
         config2.setPaths(["classpath:other"])
         config2.setMapping("/other/**")
-        DefaultStaticResourceResolver resolver = new DefaultStaticResourceResolver([config1, config2])
+        StaticResourceResolver resolver = new StaticResourceResolver([config1, config2])
 
         when:
         URL url = resolver.resolve("/").get()
